requirements.txt
streamlit
requests
Pillow
st_social_media_links
